Cambiar tipo de dato de retorno de promedio

+++ include/stats.h
@@ -1,14 +1,8 @@
 #ifndef STATS_H
 #define STATS_H
 
-/*
- * stats.h
- * Interfaz del módulo de estadísticas para la práctica de bibliotecas.
- * IE-0117 - I-2026
- */
-
 int   suma    (int *arr, int n);
-float promedio(int *arr, int n);
+int   promedio(int *arr, int n);
 int   minimo  (int *arr, int n);
 int   maximo  (int *arr, int n);
